+++ src/models/GetOrderResponseFulfillmentsInner.ts
@@ -74,6 +74,12 @@ export interface GetOrderResponseFulfillmentsInner {
    * @type {string}
    * @memberof GetOrderResponseFulfillmentsInner
    */
+  trackingStatus: string | null;
+  /**
+   *
+   * @type {string}
+   * @memberof GetOrderResponseFulfillmentsInner
+   */
   shippedAt: string | null;
   /**
    *
@@ -120,6 +126,7 @@ export function instanceOfGetOrderResponseFulfillmentsInner(
   if (!("prescriptionId" in value) || value["prescriptionId"] === undefined) return false;
   if (!("status" in value) || value["status"] === undefined) return false;
   if (!("trackingNumber" in value) || value["trackingNumber"] === undefined) return false;
+  if (!("trackingStatus" in value) || value["trackingStatus"] === undefined) return false;
   if (!("shippedAt" in value) || value["shippedAt"] === undefined) return false;
   if (!("deliveredAt" in value) || value["deliveredAt"] === undefined) return false;
   if (!("estimatedDeliveryAt" in value) || value["estimatedDeliveryAt"] === undefined) return false;
@@ -150,6 +157,7 @@ export function GetOrderResponseFulfillmentsInnerFromJSONTyped(
     prescriptionId: json["prescriptionId"],
     status: json["status"],
     trackingNumber: json["trackingNumber"],
+    trackingStatus: json["trackingStatus"],
     shippedAt: json["shippedAt"],
     deliveredAt: json["deliveredAt"],
     estimatedDeliveryAt: json["estimatedDeliveryAt"],
@@ -181,6 +189,7 @@ export function GetOrderResponseFulfillmentsInnerToJSONTyped(
     prescriptionId: value["prescriptionId"],
     status: value["status"],
     trackingNumber: value["trackingNumber"],
+    trackingStatus: value["trackingStatus"],
     shippedAt: value["shippedAt"],
     deliveredAt: value["deliveredAt"],
     estimatedDeliveryAt: value["estimatedDeliveryAt"],